code: change SPI mode
On _some_ boards, mode 2 intermittently fails; the display won't
initialize and nothing but black is displayed. This happens from 5-75% of
the time on _some_ boards.

The diagrams in the datasheet shows that mode 0 or mode 3 is correct. Many
code examples for ST7789 use mode 0, but reading the diagrams in the
datasheet looks like mode 3 to me, at least in the diagrams that clearly
show the idle clock state; many are unclear. There are many references to
using mode 3 in forums and bug reports on the Internet.

With this change, I've booted the most problematic board 400 times without
issue, with some reprogramming to use mode 2 in the middle which
reproduced the problem after 1-20 boots each time.